Task6/project/requirements.txt
# server
fastapi[all]==0.74.1


# database
databases[postgresql]==0.5.5
aerich==0.6.3
tortoise.orm[asyncpg]==0.19.0
